+++ gw2buildutil/__init__.py
@@ -1,1 +1,1 @@
-from . import api, build, defnfile, compositions
+from . import api, build, defnfile, buildtemplate, compositions